Revert "properties"
This reverts commit 4219b573580252918fad05fa0c6a6037dc8bd54e.

+++ src/main/java/com/javier/srvice/security/aplication/UserService.java
@@ -69,11 +69,7 @@ public class UserService implements UserServicePort {
     public User setProfileImage(Integer idFile, String emailAuth) throws Exception {
         User user = userRepositoryJpa.findByEmail(emailAuth).orElseThrow(() -> new Exception("That user does not exists"));
         File file = fileRepositoryJpa.findById(idFile).orElseThrow(() -> new Exception("That file does not exists"));
-        if(user.getImage()!=null) {
-            user.setImage(null);
-            userRepositoryJpa.save(user);
-            fileStoragePort.deleteFile(user.getImage().getId());
-        }
+        if(user.getImage()!=null) fileStoragePort.deleteFile(user.getImage().getId());
         user.setImage(file);
         userRepositoryJpa.save(user);
         return user;
